+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 ## Unreleased
 - Fix url used for fetching streams
 - Return `is_end_sequence` on stream fetch
+- Make `transform_tag` optional on `create bucket`
 - Retry `put_emails` requests
-
 
 ## v0.20.0